+++ src/utilities/NotificationLog.php
@@ -43,7 +43,7 @@ class NotificationLog extends Utility
     /**
      * @inheritdoc
      */
-    public static function iconPath(): ?string
+    public static function icon(): ?string
     {
         // Set the icon mask path
         $iconPath = Craft::getAlias('@vendor/doublesecretagency/craft-notifier/src/icon-mask.svg');